+++ R/ttl_write.R
@@ -8,28 +8,66 @@
 #' @param overwrite If the file exists, overwrite it? Defaults to \code{TRUE}.
 #' @return A text file with the prefix and the observation serialisations.
 #' @examples
-#' testtdf <- data.frame (s = c("eg:o1", "eg:01", "eg:02"),
-#'                        p = c("a", "eg-var:", "eg-var"),
-#'                        o = c("qb:Observation",
-#'                              "\"1\"^^<xs:decimal>",
-#'                              "\"2\"^^<xs:decimal>"))
+#' tdf <- data.frame (s = c("eg:01","eg:02",  "eg:01", "eg:02", "eg:01" ),
+#'                    p = c("a", "a", "eg-var:", "eg-var:", "rdfs:label"),
+#'                    o = c("qb:Observation",
+#'                          "qb:Observation",
+#'                          "\"1\"^^<xs:decimal>",
+#'                          "\"2\"^^<xs:decimal>", 
+#'                          '"Example observation"'))
 #'
 #' examplefile <- file.path(tempdir(), "ttl_dataset_write.ttl")
 #'
-#' dataset_ttl_write(tdf=testtdf, file_path = examplefile)
+#' dataset_ttl_write(tdf=tdf, file_path = examplefile)
 #'
 #' readLines(examplefile)
 #' @export
-#' 
-ttl_write <- function(tdf, ttl_namespace, file_path, overwrite=TRUE) {
+ 
+ttl_write <- function(tdf, ttl_namespace=NULL, file_path, overwrite=TRUE) {
   
-  getdata <- function() {
-    e <- new.env()
-    name <- utils::data("dataset_namespace", package = "dataset", envir = e)[1]
-    e[[name]]
+  if (is.null(ttl_namespace)) {
+    ## load dataset_namespace
+    default_namespace <- getdata("dataset_namespace", "dataset")
+    default_namespace <- default_namespace[
+      which(default_namespace$prefix %in% c("rdf:", "rdfs:", "owl:",
+                                            "qb:", "dcat:", "xsd:")),]
+    
+    ttl_namespace <- default_namespace
   }
   
-  dataset_namespace <- getdata()
+  ## validate dataset
+  validate_tdf(tdf)
+  
+  ## use subject predicate, object names
+  names(tdf) <- c("s", "p", "o")
+  
+  ## If no file_path is given, create a temporary file
+  if (is.null(file_path)) file_path <- file.path(tempdir(), "tmp.ttl")
+  
+  if (overwrite) {
+    reset_text_file(file_path=file_path)
+  } 
+  
+  ttl_prefix_write(ttl_namespace = ttl_namespace,
+                   file_path=file_path,
+                   overwrite=overwrite)
 
-  dataset::dataset_ttl_write(tdf, ttl_namespace=dataset_namespace, file_path, overwrite)
+  ttl_statements_write(tdf, file_path, overwrite=FALSE)
+}
+
+#' @title Validate tdf
+#' @description
+#' The tdf object must be inherited from data.frame and have exactly three columns.
+#' @keywords internal
+validate_tdf <- function(tdf) {
+  
+  if ( ! inherits(tdf, "data.frame") ) {
+    stop("ttl_dataset_write(): tdf must be a data.frame (inherited) object")
+  }
+  
+  if ( ncol(tdf) != 3 ) {
+    stop("ttl_dataset_write(): tdf must have exactly three columns.")
+  }
+  
+  TRUE
 }