+++ NEWS.md
@@ -1,3 +1,4 @@
-# tuRtle 0.1.0
+# tuRtle 0.1.1
 
-* Initial development version on Zenodo and GitHub.
+* 2024-01-28. Initial development version 0.1.0 on Zenodo and GitHub.
+* 2024-06-11. Version 0.1.11 has separate functions for writing a namespace prefix and statements, and has extensive unit tests.